Bare Bones
I started with a simple template that I adapted from a similar project online, mostly because it had a way to assign card faces to each card created. The rules are very basic and this version doesn't run anything as is, will add functionality to run a player vs computer game that only allows the player to hit, stay, and reset the game.