Form builds and runs. but not quite correctly
I have all of the methods now filled in, and have assigned the player constructed in the Table class a starting balance of 4000, as otherwise it would start blank if I tried to do so in the Player class. The Form loads, but will not show cards that should already be dealt and throws an exception when any button is pressed after a bet is made. The betting increment and clearing buttons work as intended, as do the new and exit game buttons.